[Fix] Escala y Promocion

+++ TECAir_API/Models/WebOutput/PromotionOutput.cs
@@ -2,20 +2,20 @@
 {
     public class PromotionOutput
     {
-        public decimal NoPromocion { get; set; }
-        public decimal Porcentaje { get; set; }
-        public decimal Periodo { get; set; }
+        public int NoPromocion { get; set; }
+        public int Porcentaje { get; set; }
+        public int Periodo { get; set; }
         public string Url { get; set; }
         public string PDia { get; set; }
         public string PMes { get; set; }
         public string PAno { get; set; }
-        public decimal NoVuelo { get; set; }
+        public int NoVuelo { get; set; }
 
         public PromotionOutput()
         {
         }
 
-        public PromotionOutput(decimal noPromocion, decimal porcentaje, decimal periodo, string url, string pDia, string pMes, string pAno, decimal noVuelo)
+        public PromotionOutput(int noPromocion, int porcentaje, int periodo, string url, string pDia, string pMes, string pAno, int noVuelo)
         {
             NoPromocion = noPromocion;
             Porcentaje = porcentaje;
